requirements.txt
requests
python-dotenv
influxdb-client
pandas
numpy
supabase
